renamed static array header to match updates
